% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addq.R
\name{addq}
\alias{addq}
\title{Wrapper to quote name}
\usage{
addq(x)
}
\arguments{
\item{x}{A string}
}
\value{
\code{x} between quotes, as it must be to be supported by regex
}
\description{
Wrapper to quote name
}
