% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_simplelasso.R
\name{big.simplelasso}
\alias{big.simplelasso}
\title{Perform LASSO to select best features for big.matrix objects}
\usage{
big.simplelasso(X, yvar = "incidence", labelvar = c("cancer", "age",
  "Country_Transco", "year", "area.x", "area.y"), crossvalidation = T,
  nfolds = 10, returnplot = T, ncores = 1)
}
\arguments{
\item{X}{A \code{big.matrix} object}

\item{yvar}{Column name for the explained variable}

\item{labelvar}{Additional variables providing information but
that should not be used as explanatory variables}

\item{crossvalidation}{Should cross-validation be performed? TRUE or FALSE}

\item{nfolds}{Number of folds for cross-validation. Ignored if
\code{crossvalidation = T}}

\item{ncores}{Number of cores to used for computation. If \code{ncores}>1,
parallel processing is used}
}
\value{
A list of three elements.

\code{output$model} returns the model. \cr

\code{output$plot} returns a plot. If \code{crossvalidation = F},
coefficients values when $\lambda$ penalization term
evolves is represented.
If \code{crossvalidation = T}, the RMSE is represented with respect to the
number of variables with non-zero weight \cr

\code{output$coeff} returns the coefficient values returned by the LASSO
(or the coefficients of the RMSE-minimizing model if \code{crossvalidation = T})
}
\description{
\code{big.simplelasso} estimates a model using LASSO and
returns a sparse structure.
Cross-validation can be used to select best covariates
combination. The dataframe must be a C++ pointer
}
