% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cleanimport.R
\name{check.clean.import}
\alias{check.clean.import}
\title{Check if dataframe import is clean}
\usage{
check.clean.import(path = getwd(), filename = "training.csv",
  training = T)
}
\arguments{
\item{path}{Directory where the csv file can be found}

\item{filename}{Filename with extension}

\item{training}{Boolean indicating whether we import
training dataset or other datasets (formatting differ)}
}
\description{
\code{check.clean.import} imports a file
using \code{readr::read_csv} and \code{utils::read.csv}
routines and compare whether they are equal or
not. If they differ, a more precise look is made by
testing near equality (\code{base::all.equal}) and a
diagnosis is returned
}
\examples{
\dontrun{
# To run following commands, you should
# define a datadir path where all files are located
diag1 <- check.clean.import(path = datadir)
diag2 <- check.clean.import(path = datadir, filename = "WorldBank_Data.csv",training = F)
diag3 <- check.clean.import(path = datadir, filename = "Ilostat_Data.csv",training = F)
diag4 <- check.clean.import(path = datadir, filename = "Faostat_Data.csv",training = F)}
}
