% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_emptycolumns.R
\name{check.emptycolumn}
\alias{check.emptycolumn}
\title{Count share of missing values in a variable}
\usage{
check.emptycolumn(data, proportion = F)
}
\arguments{
\item{data}{Dataframe to inspect}

\item{proportion}{Boolean indicating whether we
should return proportion of missing values or a
boolean vector indicating whether all values are
missing or not}
}
\value{
A named vector (from \code{data} colnames)
of missing values proportion or TRUE/FALSE vector
}
\description{
Count share of missing values in a variable
}
\examples{
df <- data.frame(x1 = rnorm(10),x2 = rnorm(10))
df$x3 <- rep(NA,10)
df$x1[df$x1<0] <- NA
check.emptycolumn(df)
check.emptycolumn(df,proportion = T)
}
