% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_character.R
\name{convert.character}
\alias{convert.character}
\title{Convert all character variables to numeric}
\usage{
convert.character(df)
}
\arguments{
\item{df}{Dataframe}
}
\value{
A list of two elements. The first one (\code{data}),
is the modified dataframe. The second one (\code{correspondance})
is a list of all transformation that have been applied
}
\description{
\code{convert.character} can be used to detect
and convert all character vectors within a
dataframe to numeric. A correspondance table of all
performed transmutation is also returned
}
