% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlag.R
\name{createlag}
\alias{createlag}
\title{Transform dataframe by creating multiple lagged variables}
\usage{
createlag(df, groupingvar = c("sex", "age", "Country_Transco"), k = 1,
  labelvar = c("year", "Zonier"))
}
\arguments{
\item{df}{Dataframe whose variables must be lagged}

\item{groupingvar}{Variables used to define group}

\item{k}{Maximum lag that must be computed. If k>1, lags 1:k are
computed}

\item{labelvar}{Variables that should not be lagged}
}
\value{
A dataframe with lagged observations by groups
}
\description{
Given a dataframe, \code{createlag} returns a dataframe with lagged
variables. Lag function is applied by groups. Additional variables
that should not be lagged can be provided
}
