% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_dummies.R
\name{expand.dummies}
\alias{expand.dummies}
\title{Transfor factor variables with a set of dummies}
\usage{
expand.dummies(df, labelvar = "sex")
}
\arguments{
\item{df}{Dataframe}

\item{labelvar}{Factor column name that must be excluded from
the transformation}
}
\value{
Dataframe with factor variables converted as a set of
dummy variables
}
\description{
\code{expand.dummies} applies, for all factor
variables, an expansion of the variable in wide format
by setting several dummy columns for all possible values of
the variable
}
