% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FE_lasso.R
\name{FE.lasso}
\alias{FE.lasso}
\title{Perform LASSO for feature selection and create linear
regression using selected features}
\usage{
FE.lasso(df, crossvalidation = T, nfolds = 10, include.intercept = TRUE,
  relabel = F)
}
\arguments{
\item{df}{Dataframe}

\item{crossvalidation}{Should we perform cross-validation}

\item{nfolds}{Number of folds for cross validation.
Ignored if \code{crossvalidation = F}. If \code{crossvalidation}
equals the number of observations, a leave-one-out cross validation
is performed}

\item{include.intercept}{Should LASSO model include a constant?}
}
\value{
Linear model fitted with selected features
}
\description{
Perform LASSO for feature selection and create linear
regression using selected features
}
