% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dataframe.R
\name{fill_dataframe}
\alias{fill_dataframe}
\title{Filling missing intermediate year vector}
\usage{
fill_dataframe(df, year.var = "year", label.var = c("ref_area", "area",
  "Country_Transco", "Zonier"))
}
\arguments{
\item{df}{Dataframe with unbalanced year vector}

\item{year.var}{Column name of the year variable}

\item{label.var}{Columns that should be filled by
constant values and not by NAs}
}
\value{
Dataframe with intermediate years added filled
by NAs
}
\description{
Filling missing intermediate year vector
}
