% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_FAO.R
\name{import_FAO}
\alias{import_FAO}
\alias{import_WB}
\alias{import_ILO}
\alias{import_training}
\title{Importing data from Epidemium}
\usage{
import_FAO(path = getwd(), filename = "Faostat_Data.csv", fromURL = FALSE,
  url = "http://qa.epidemium.cc/data/epidemiology_dataset/fao_data/Faostat_Data.csv",
  colstokeep = 1:10, fill.Ethiopia = T)

import_WB(path = getwd(), filename = "WorldBank_Data.csv",
  fromURL = FALSE,
  url = "http://qa.epidemium.cc/data/epidemiology_dataset/world_bank_data/WorldBank_Data.csv",
  colstokeep = 1:10)

import_ILO(path = getwd(), filename = "Ilostat_Data.csv", fromURL = FALSE,
  url = "http://qa.epidemium.cc/data/epidemiology_dataset/ilo_data/Ilostat_Data.csv",
  colstokeep = 1:10)

import_training(path = getwd(), filename = "training.csv",
  fromURL = FALSE,
  url = "http://qa.epidemium.cc/data/incidence_dataset/2017-10-10_dataset_core/training.txt",
  colstokeep = 1:10)
}
\arguments{
\item{path}{The path where the file can be found (ignored if file
is loaded from the web)}

\item{filename}{Name of the file with extension (csv required)}

\item{fromURL}{Boolean indicating whether the file should be
loaded from computer (\code{fromURL = FALSE}) or internet
((\code{fromURL = TRUE}))}

\item{url}{The url where the file can be found (ignored if \code{fromURL = FALSE})}

\item{colstokeep}{A sequence indicating which columns to keep. If \code{NULL}
the whole dataframe is returned}

\item{fill.Ethiopia}{Specific to \code{import_FAO}. Boolean indicating whether we
should fill Ethiopia data with Ethiopia PDR data}
}
\value{
A tibble with requested data
}
\description{
These functions allow you to import FAO, World Bank, ILO data
and training data
from your computer
or directly from the internet
}
\section{Warning}{

Downloading from URL can be very long, it is recommended to load from computer
}

\examples{
\dontrun{
df_FAO <- import_FAO(getwd(), colstokeep = 1:10)
df_WB  <- import_WB(fromURL = T, colstokeep = 1:6)
df_ILO <- import_ILO(paste0(getwd(),"/other"))
df_training <- import_training(path = datadir, colstokeep = NULL)}
}
