% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_data.R
\name{interpolate_group}
\alias{interpolate_group}
\title{Given a grouped dataframe and a line, interpolate}
\usage{
interpolate_group(missingdata, df, i = 1, year.var = "year")
}
\arguments{
\item{missingdata}{A dataframe listing groups and variables
that must be changed}

\item{i}{Line index of missingdata that will be looked}
}
\value{
A list with several components: \code{lines} and \code{cols}
return the lines and columns concerned by interpolation, \code{after}
returns the relevant lines and columns interpolated while \code{change}
stores as a dummy whether data have been interpolated
}
\description{
Given a grouped dataframe and a line, interpolate
}
\details{
Intermediate function, do not use
}
\examples{
# See \\code{interpolate_data} source code
}
