% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_variables.R
\name{label_variables}
\alias{label_variables}
\title{Use variables labels in a regression}
\usage{
label_variables(model)
}
\arguments{
\item{model}{An object of class \code{lm}
or \code{big.lm}}
}
\value{
Same object with modified variable names
}
\description{
Use variables labels in a regression
}
