% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lassocoeff.R
\name{lasso.coeff}
\alias{lasso.coeff}
\title{Recover coefficients from a LASSO model}
\usage{
lasso.coeff(model)
}
\arguments{
\item{model}{A \code{glmnet} model}
}
\value{
A dataframe with coefficients values and
associated variable name by descending order
}
\description{
This function arranges the output of glmnet
models to easily access coefficient values
}
\section{Warnings}{

Coefficients are ranked by descending order in absolute
value. Confidence intervals are not returned since there
is no obvious distribution for parameter values in LASSO
}

