% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchkeys.R
\name{matchkeys}
\alias{matchkeys}
\title{Match country name with its unique key}
\usage{
matchkeys(df_data, df_keys, origin = "FaoStat", countrycol = "area")
}
\arguments{
\item{df_data}{Dataframe with a country vector}

\item{df_keys}{Dataframe where countries are associated to
a unique key}

\item{origin}{In \code{df_keys}, which data source should
we look at?}

\item{countrycol}{Column name of the variable where countries
can be found in \code{df_data}}
}
\value{
A dataframe with unique id keys
}
\description{
Associate to a country its unique identifier necessary
to merge dataframes
}
\examples{
\dontrun{
codes <- readxl::read_excel(path = paste0(datadir,"/creation zonier.xls"),
sheet = "Transco_Country")
# MAPPING IDs TO THEIR VALUES
df_FAO = matchkeys(df_FAO,codes)
df_WB = matchkeys(df_WB,codes,origin = "World Bank")
df_ILO = matchkeys(df_ILO,codes, origin = "IloStat")}
}
