% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_FElasso.R
\name{model.FElasso}
\alias{model.FElasso}
\title{LASSO feature selection and linear regression by group}
\usage{
model.FElasso(df, groupingvar = NULL, crossvalidation = T, nfolds = 10,
  include.intercept = TRUE)
}
\arguments{
\item{df}{Dataframe}

\item{groupingvar}{Column names of variables used for group. If
\code{groupingvar = NULL}, observations are ungrouped() and pooled
estimation is performed.}

\item{nfolds}{Number of folds for cross validation.
Ignored if \code{crossvalidation = F}. If \code{crossvalidation}
equals the number of observations, a leave-one-out cross validation
is performed}

\item{include.intercept}{Should LASSO model include a constant?}
}
\description{
LASSO feature selection and linear regression by group
}
