% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenCancer.R
\docType{package}
\name{OpenCancer}
\alias{OpenCancer}
\alias{OpenCancer-package}
\title{OpenCancer: A statistical package for understanding colon cancer}
\description{
This package provides several convenient functions to estimate the
factors that can cause colon cancer
}
\details{
To learn more about the package, start with the
vignettes: \code{browseVignettes(package = "OpenCancer")}
}
