% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_FElasso.R
\name{predict.FElasso}
\alias{predict.FElasso}
\title{Prediction after a model.FElasso routine}
\usage{
\method{predict}{FElasso}(model, newdata, grouped = FALSE)
}
\arguments{
\item{model}{A nested dataframe or and lm model}

\item{newdata}{Dataframe}

\item{grouped}{Boolean indicating if newdata should be
grouped}
}
\description{
Prediction after a model.FElasso routine
}
