% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recover_data.R
\name{recover_data}
\alias{recover_data}
\title{Perform feature selection on pointers and
import data back in R memory}
\usage{
recover_data(X, yvar = "incidence", labelvar = c("cancer", "age",
  "Country_Transco", "year", "area.x", "area.y"), crossvalidation = T,
  nfolds = 10, returnplot = F, ncores = 1)
}
\arguments{
\item{X}{A \code{big.matrix} object}

\item{yvar}{Name of the variable we want to
explain in \code{X}}

\item{labelvar}{Variable names of columns excluded from
the set of covariates in LASSO. All variables
included in \code{labelvar} will be added to the set of
selected features imported back in memory}

\item{crossvalidation}{Should we perform a cross-validated
LASSO ? \code{TRUE} or \code{FALSE}}

\item{nfolds}{Number of folds for cross-validation. Ignored if
\code{crossvalidation = FALSE}. If \code{nfolds} equals the number
of observations, leave-one-out cross validation is performed}

\item{returnplot}{Should we return a plot of the relationship
between quadratic risk and $\lambda$ parameter? \code{TRUE} or
\code{FALSE}}

\item{ncores}{Number of cores used for computations.
If \code{ncores>1}, parallel processing is used}
}
\description{
\code{recover_data} has been designed to
easily perform feature selection by LASSO on
pointers and import selected features as well as
additional relevant variables back in the memory
}
